requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.7
python-dateutil>=2.8
pytest>=7.0
jsonschema>=4.0
